MovementEngine is now threaded!
The majority of the pathfinding-calculation now runs in a seperate thread. It is still not opitmal, as the path-optimizer still runs in the main thread, however I was unable to find a solution for this.
Anyway, performance is now much improved (Tested with 30 or more units at once).
Please have a look at it and optimize it if possible.

Problem: Some units get simply stuck at some edge of a building and do not continue to move. (Problem was there before, but needs to be fixed).